Unit Tests finished and Web Server re architectured

+++ C#/Gamify.Sdk.Tests/DataTests/RepositoryTests.cs
@@ -2,13 +2,21 @@
 using Gamify.Sdk.Data.Configuration;
 using Gamify.Sdk.Data.Entities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
 
 namespace Gamify.Sdk.Tests.DataTests
 {
     [TestClass]
     public abstract class RepositoryTests<T> where T : MongoEntity
     {
+        private readonly bool cleanDbWhenFinishes;
+
         protected IRepository<T> testRepository;
+
+        protected RepositoryTests(bool cleanDbWhenFinishes = true)
+        {
+            this.cleanDbWhenFinishes = cleanDbWhenFinishes;
+        }
 
         [TestInitialize]
         public void Initialize()
@@ -21,7 +29,17 @@ namespace Gamify.Sdk.Tests.DataTests
         [TestCleanup]
         public void Cleanup()
         {
-            this.testRepository.DeleteAll();
+            if (this.cleanDbWhenFinishes)
+            {
+                this.testRepository.DeleteAll();
+            }
+        }
+
+        protected static string GetUniqueName(string name = null)
+        {
+            var uniqueName = Guid.NewGuid().ToString();
+
+            return string.IsNullOrEmpty(name) ? uniqueName : string.Concat(name, "-", uniqueName);
         }
     }
 }
